oauthd_v.sh -> verbose server

+++ oauthd_v.sh
@@ -8,17 +8,17 @@ oauthdurl=`node -e 'var c=require("./lib/config");console.log("http://127.0.0.1:
 
 case "$1" in
 start)
-	if curl -o /dev/null --silent --fail "$oauthdurl"; then
+	if curl -fail "$oauthdurl"; then
 		echo -n "Stopping OAuth daemon..."
-		forever stop lib/oauthd.js >/dev/null 2>/dev/null
+		forever stop lib/oauthd.js
 		echo " OK"
 	fi
 	echo -n "Starting OAuth daemon."
-	forever --minUptime 1000 --spinSleepTime 1000 -a -l forever.log -o out.log -e err.log start lib/oauthd.js
+	forever --minUptime 1000 --spinSleepTime 1000 -a -f start lib/oauthd.js
 	for i in {1..10}
 	do
 		sleep 1
-	if curl --fail "$oauthdurl"; then
+	if curl -fail "$oauthdurl"; then
 		echo " OK"
 		node -e 'var c=require("./lib/config");console.log("Admin interface at "+c.host_url+c.base+"/admin")'
 		exit 0
@@ -29,16 +29,16 @@ start)
 	echo " Failed"
 ;;
 status)
-	if curl -o /dev/null --silent --fail "$oauthdurl"; then
+	if curl -fail "$oauthdurl"; then
 		echo "OAuth daemon is running"
 	else
 		echo "OAuth daemon is stopped"
 	fi
 ;;
 stop)
-	if curl -o /dev/null --silent --fail "$oauthdurl"; then
+	if curl -fail "$oauthdurl"; then
 		echo -n "Stopping OAuth daemon..."
-		forever stop lib/oauthd.js >/dev/null 2>/dev/null
+		forever stop lib/oauthd.js
 		echo " OK"
 	else
 		echo "OAuth daemon is already stopped"
@@ -49,10 +49,10 @@ restart)
 	$0 start
 ;;
 startsync)
-	forever --minUptime 1000 --spinSleepTime 1000 -a -l forever.log -o out.log -e err.log lib/oauthd.js
+	forever --minUptime 1000 --spinSleepTime 1000 -a -f lib/oauthd.js
 ;;
 config)
-	if [ ! -f ./config.local.js ]; then
+	if [ ! -a -f ./config.local.js ]; then
 		cp ./config.js ./config.local.js
 	fi
 	if [ "$EDITOR" != "" ]; then
